requirements.txt
python-dotenv
chromadb
google-generativeai
streamlit
youtube-dl
youtube-transcript-api
numpy
pandas
pillow
matplotlib
langchain
transformers